Support sparse images via virtio-blk
This adds support for a custom sparse image format ("SILT").

A SILT file looks like:

  4 bytes: S L T 0x01 (magic numbers)
  4 bytes: n_sectors (uint32, big endian)
  n_sectors * 4 bytes: Lookup Table
  rest of file: data

The lookup table maps to indexes in the data segment. It's prefilled
with 0xff, and a LUT entry of 0xffffffff is used to indicate a sector is
full of zeroes.

To write 0x01 to the 2048'th byte of the disk, with a 1024-byte sector
size, for example:

  1. 2048 / 1024 == 2, 2048 % 1024 == 0, meaning we write into byte zero
    of segment 2.
  2. Read the four bytes at LUT[2 * 4]. This is the offset into the data
    at which this 1k sector is located.
  3. Write out a 1k chunk at that index into the data with 0x01 set.

This is heavily based on a related branch at:
https://github.com/dunkelstern/xhyve/tree/sparse-disk-image

This implementation differs in a few ways:

* No support for image splitting
* LUT and data unified in one file
* magic numbers and sector count encoded in file, verified on load
* bugs fixed